AWSSYSTEM/requirements.txt
mongoengine
paho-mqtt
six 
pytz